+++ src/adapter/driver/DiskConverter.java
@@ -343,35 +343,4 @@ public class DiskConverter
         diskAccess.seek(atRef - 2 + EntryLength);   //Vai para o final do DEntry (should)
         return dEntry;
     }
-    /*
-    public byte[] ReadBytes(int numberOfBytes)
-    throws IOException
-    {
-        byte[] byteArray = new byte[numberOfBytes];
-        diskAccess.read(byteArray, 0, numberOfBytes);
-        return byteArray;
-    }
-
-    public long ByteArrayToLong(byte[] byteArray)
-    {
-        ByteBuffer wrapper = ByteBuffer.wrap(byteArray);
-        return wrapper.getLong();
-    }
-    public int ByteArrayToInt(byte[] byteArray)
-    {
-        ByteBuffer wrapper = ByteBuffer.wrap(byteArray);
-        return wrapper.getInt();
-    }
-    public int ByteArrayToUnsignedShort(byte[] byteArray)
-    {
-        ByteBuffer wrapper = ByteBuffer.wrap(byteArray);
-        short ShortValue = wrapper.getShort();
-        return ShortValue >= 0? ShortValue : 0x10000 + ShortValue;
-    }
-    public short ByteArrayToShort(byte[] byteArray)
-    {
-        ByteBuffer wrapper = ByteBuffer.wrap(byteArray);
-        return wrapper.getShort();
-    }
-    */
 }
